requirements.txt
agency-swarm==0.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argcomplete==3.5.1
asttokens==3.0.0
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.10.0
bleach==6.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
datamodel-code-generator==0.26.1
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
email_validator==2.2.0
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
fsspec==2024.10.0
genson==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflect==5.6.2
ipython==8.12.3
isort==5.13.2
jedi==0.19.2
Jinja2==3.1.4
jiter==0.7.1
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
nbclient==0.10.1
nbconvert==7.16.4
nbformat==5.10.4
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.54.4
openai-whisper==20240930
ordered-set==4.1.0
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
platformdirs==4.3.6
prompt_toolkit==3.0.48
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pyobjc==10.3.2
pyobjc-core==10.3.2
pyobjc-framework-Accessibility==10.3.2
pyobjc-framework-Accounts==10.3.2
pyobjc-framework-AddressBook==10.3.2
pyobjc-framework-AdServices==10.3.2
pyobjc-framework-AdSupport==10.3.2
pyobjc-framework-AppleScriptKit==10.3.2
pyobjc-framework-AppleScriptObjC==10.3.2
pyobjc-framework-ApplicationServices==10.3.2
pyobjc-framework-AppTrackingTransparency==10.3.2
pyobjc-framework-AudioVideoBridging==10.3.2
pyobjc-framework-AuthenticationServices==10.3.2
pyobjc-framework-AutomaticAssessmentConfiguration==10.3.2
pyobjc-framework-Automator==10.3.2
pyobjc-framework-AVFoundation==10.3.2
pyobjc-framework-AVKit==10.3.2
pyobjc-framework-AVRouting==10.3.2
pyobjc-framework-BackgroundAssets==10.3.2
pyobjc-framework-BrowserEngineKit==10.3.2
pyobjc-framework-BusinessChat==10.3.2
pyobjc-framework-CalendarStore==10.3.2
pyobjc-framework-CallKit==10.3.2
pyobjc-framework-CFNetwork==10.3.2
pyobjc-framework-Cinematic==10.3.2
pyobjc-framework-ClassKit==10.3.2
pyobjc-framework-CloudKit==10.3.2
pyobjc-framework-Cocoa==10.3.2
pyobjc-framework-Collaboration==10.3.2
pyobjc-framework-ColorSync==10.3.2
pyobjc-framework-Contacts==10.3.2
pyobjc-framework-ContactsUI==10.3.2
pyobjc-framework-CoreAudio==10.3.2
pyobjc-framework-CoreAudioKit==10.3.2
pyobjc-framework-CoreBluetooth==10.3.2
pyobjc-framework-CoreData==10.3.2
pyobjc-framework-CoreHaptics==10.3.2
pyobjc-framework-CoreLocation==10.3.2
pyobjc-framework-CoreMedia==10.3.2
pyobjc-framework-CoreMediaIO==10.3.2
pyobjc-framework-CoreMIDI==10.3.2
pyobjc-framework-CoreML==10.3.2
pyobjc-framework-CoreMotion==10.3.2
pyobjc-framework-CoreServices==10.3.2
pyobjc-framework-CoreSpotlight==10.3.2
pyobjc-framework-CoreText==10.3.2
pyobjc-framework-CoreWLAN==10.3.2
pyobjc-framework-CryptoTokenKit==10.3.2
pyobjc-framework-DataDetection==10.3.2
pyobjc-framework-DeviceCheck==10.3.2
pyobjc-framework-DictionaryServices==10.3.2
pyobjc-framework-DiscRecording==10.3.2
pyobjc-framework-DiscRecordingUI==10.3.2
pyobjc-framework-DiskArbitration==10.3.2
pyobjc-framework-DVDPlayback==10.3.2
pyobjc-framework-EventKit==10.3.2
pyobjc-framework-ExceptionHandling==10.3.2
pyobjc-framework-ExecutionPolicy==10.3.2
pyobjc-framework-ExtensionKit==10.3.2
pyobjc-framework-ExternalAccessory==10.3.2
pyobjc-framework-FileProvider==10.3.2
pyobjc-framework-FileProviderUI==10.3.2
pyobjc-framework-FinderSync==10.3.2
pyobjc-framework-FSEvents==10.3.2
pyobjc-framework-GameCenter==10.3.2
pyobjc-framework-GameController==10.3.2
pyobjc-framework-GameKit==10.3.2
pyobjc-framework-GameplayKit==10.3.2
pyobjc-framework-HealthKit==10.3.2
pyobjc-framework-ImageCaptureCore==10.3.2
pyobjc-framework-InputMethodKit==10.3.2
pyobjc-framework-InstallerPlugins==10.3.2
pyobjc-framework-InstantMessage==10.3.2
pyobjc-framework-Intents==10.3.2
pyobjc-framework-IntentsUI==10.3.2
pyobjc-framework-IOBluetooth==10.3.2
pyobjc-framework-IOBluetoothUI==10.3.2
pyobjc-framework-IOSurface==10.3.2
pyobjc-framework-iTunesLibrary==10.3.2
pyobjc-framework-KernelManagement==10.3.2
pyobjc-framework-LatentSemanticMapping==10.3.2
pyobjc-framework-LaunchServices==10.3.2
pyobjc-framework-libdispatch==10.3.2
pyobjc-framework-libxpc==10.3.2
pyobjc-framework-LinkPresentation==10.3.2
pyobjc-framework-LocalAuthentication==10.3.2
pyobjc-framework-LocalAuthenticationEmbeddedUI==10.3.2
pyobjc-framework-MailKit==10.3.2
pyobjc-framework-MapKit==10.3.2
pyobjc-framework-MediaAccessibility==10.3.2
pyobjc-framework-MediaLibrary==10.3.2
pyobjc-framework-MediaPlayer==10.3.2
pyobjc-framework-MediaToolbox==10.3.2
pyobjc-framework-Metal==10.3.2
pyobjc-framework-MetalFX==10.3.2
pyobjc-framework-MetalKit==10.3.2
pyobjc-framework-MetalPerformanceShaders==10.3.2
pyobjc-framework-MetalPerformanceShadersGraph==10.3.2
pyobjc-framework-MetricKit==10.3.2
pyobjc-framework-MLCompute==10.3.2
pyobjc-framework-ModelIO==10.3.2
pyobjc-framework-MultipeerConnectivity==10.3.2
pyobjc-framework-NaturalLanguage==10.3.2
pyobjc-framework-NetFS==10.3.2
pyobjc-framework-Network==10.3.2
pyobjc-framework-NetworkExtension==10.3.2
pyobjc-framework-NotificationCenter==10.3.2
pyobjc-framework-OpenDirectory==10.3.2
pyobjc-framework-OSAKit==10.3.2
pyobjc-framework-OSLog==10.3.2
pyobjc-framework-PassKit==10.3.2
pyobjc-framework-PencilKit==10.3.2
pyobjc-framework-PHASE==10.3.2
pyobjc-framework-Photos==10.3.2
pyobjc-framework-PhotosUI==10.3.2
pyobjc-framework-PreferencePanes==10.3.2
pyobjc-framework-PushKit==10.3.2
pyobjc-framework-Quartz==10.3.2
pyobjc-framework-QuickLookThumbnailing==10.3.2
pyobjc-framework-ReplayKit==10.3.2
pyobjc-framework-SafariServices==10.3.2
pyobjc-framework-SafetyKit==10.3.2
pyobjc-framework-SceneKit==10.3.2
pyobjc-framework-ScreenCaptureKit==10.3.2
pyobjc-framework-ScreenSaver==10.3.2
pyobjc-framework-ScreenTime==10.3.2
pyobjc-framework-ScriptingBridge==10.3.2
pyobjc-framework-SearchKit==10.3.2
pyobjc-framework-Security==10.3.2
pyobjc-framework-SecurityFoundation==10.3.2
pyobjc-framework-SecurityInterface==10.3.2
pyobjc-framework-SensitiveContentAnalysis==10.3.2
pyobjc-framework-ServiceManagement==10.3.2
pyobjc-framework-SharedWithYou==10.3.2
pyobjc-framework-SharedWithYouCore==10.3.2
pyobjc-framework-ShazamKit==10.3.2
pyobjc-framework-Social==10.3.2
pyobjc-framework-SoundAnalysis==10.3.2
pyobjc-framework-Speech==10.3.2
pyobjc-framework-SpriteKit==10.3.2
pyobjc-framework-StoreKit==10.3.2
pyobjc-framework-Symbols==10.3.2
pyobjc-framework-SyncServices==10.3.2
pyobjc-framework-SystemConfiguration==10.3.2
pyobjc-framework-SystemExtensions==10.3.2
pyobjc-framework-ThreadNetwork==10.3.2
pyobjc-framework-UniformTypeIdentifiers==10.3.2
pyobjc-framework-UserNotifications==10.3.2
pyobjc-framework-UserNotificationsUI==10.3.2
pyobjc-framework-VideoSubscriberAccount==10.3.2
pyobjc-framework-VideoToolbox==10.3.2
pyobjc-framework-Virtualization==10.3.2
pyobjc-framework-Vision==10.3.2
pyobjc-framework-WebKit==10.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.7.1
rpds-py==0.22.3
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
termcolor==2.4.0
tiktoken==0.8.0
tinycss2==1.4.0
torch==2.5.1
tornado==6.4.2
tqdm==4.67.0
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
whisper==1.1.10
whisper-at==0.5
yarg==0.1.9
